Update homepage
This is a draft of the Home Page for you to edit. This is a basic draft for HTML. The CSS page is operating in the background. If you would like to change the layout or style etc...please let me know so that I can update the CSS page.